feat: migrate ActionState and related classes to NPCProcGen for improved state management
