+++ addons/NPCProcGen/Core/States/WanderState.cs
@@ -1,0 +1,14 @@
+namespace NPCProcGen.Core.States
+{
+    public class WanderState : ActionState
+    {
+        public WanderState(ActorTag2D owner) : base(owner)
+        {
+        }
+
+        public override void Update()
+        {
+
+        }
+    }
+}
